+++ src/modules/proto_dhcp/dhcpclient.c
@@ -459,13 +459,13 @@ int main(int argc, char **argv)
 			break;
 #endif
 		case 'r':
-			if (!isdigit((int) *optarg))
+			if (!isdigit((uint8_t) *optarg))
 				usage();
 			retries = atoi(optarg);
 			if ((retries == 0) || (retries > 1000)) usage();
 			break;
 		case 't':
-			if (!isdigit((int) *optarg))
+			if (!isdigit((uint8_t) *optarg))
 				usage();
 			timeout = atof(optarg);
 			break;
@@ -530,7 +530,7 @@ int main(int argc, char **argv)
 	 *	See what kind of request we want to send.
 	 */
 	if (argc >= 3) {
-		if (!isdigit((int) argv[2][0])) {
+		if (!isdigit((uint8_t) argv[2][0])) {
 			packet_code = fr_str2int(request_types, argv[2], -2);
 			if (packet_code == -2) {
 				fprintf(stderr, "Unknown packet type: %s\n", argv[2]);
